Speed up exact MATRIX_SUMMARY: float-division bucketing + register coalescing
The exact per-cell rewrite was dominated by the per-cell output-cell bucketing.
ncu showed the matrix-summary kernel was FP64-bound (~87% FP64 pipeline) -- the
bottleneck was the double division floor((pos)/cells_per_bucket) done for every
distance, not the atomics. Two changes remove it:

1. Bucket in float instead of double. cols_per_cell / rows_per_cell are now
   stored as float in SCAMPKernelInputArgs and the division/floor runs on the
   FP32 pipe. A plain float divide (not a reciprocal multiply -- that rounds
   1/c separately and shifted ~91 boundary columns into the wrong cell) is
   accurate enough that the exact reduce_matrix verification still passes for
   every test. Positions are passed as int (tile-local, < 2^31), so the bucket
   math is int->float + one float divide.

2. Register-coalesce the per-cell accumulation: each thread keeps the running
   max for its current output cell in a register and only emits a (block-scoped)
   smem atomic when the cell changes, flushing the trailing accumulator before
   write_back. Distances almost always map to the same cell as cells span many
   rows/cols, so this collapses many atomics into one.

Also remove the now-dead per-column write_back_value MATRIX_SUMMARY branch
(write_back flushes the smem cell grid directly) and add a --profile_types
filter to test/run_tests.py so a single profile can be exercised in isolation.

RTX 3080, 1M self-join, cold sliding-window default:
  MATRIX_SUMMARY SINGLE  41.7s -> 5.3s   (master 7.3s)
  MATRIX_SUMMARY DOUBLE  51.8s -> 17.5s  (master 166.6s)
Both precisions now beat master with the exact per-cell result. Verified:
GPU sliding-window 6015/6015, GPU shfl (forced) 3943/3943, 0 failures.

+++ src/core/kernel_common.h
@@ -29,8 +29,8 @@ struct SCAMPKernelInputArgs {
   int32_t exclusion_upper;
   int32_t matrix_width;
   int32_t matrix_height;
-  double rows_per_cell;
-  double cols_per_cell;
+  float rows_per_cell;
+  float cols_per_cell;
   int64_t global_start_col;
   int64_t global_start_row;
   bool has_nan_input;
